+++ src/games/solitaire/klondike/klondikeFieldRenderer/klondikeFieldRenderer.cpp
@@ -1,25 +1,13 @@
 #include "klondikeFieldRenderer.h"
 
 namespace WGP {
-	// Constructors
-
-	klondikeFieldRenderer::klondikeFieldRenderer() :
-		_klondikeField(nullptr),
-		_cardAtlas(nullptr),
-		_klondikeAtlas(nullptr),
-		_cardsScale(1.f),
-		_cardsSpacing(0.f),
-		_columnSpacing(0.f),
-		_rowSpacing(0.f),
-		_startPosition(sf::Vector2f(OUT_OF_BOUNDS, OUT_OF_BOUNDS)),
-		_isRendering(false)
-	{
-
-	}
-
-	klondikeFieldRenderer::klondikeFieldRenderer(sf::Texture& cardAtlas, sf::Texture& klondikeAtlas, klondikeField* field, float scale, float cardsSpacing, float colSpacing, float rowSpacing, sf::Vector2f startPos) :
+	// Constructor
+	klondikeFieldRenderer::klondikeFieldRenderer(sf::Texture& cardAtlas, sf::Texture& klondikeAtlas, klondikeField* field, const sf::Font& font, float scale, float cardsSpacing, float colSpacing, float rowSpacing, sf::Vector2f startPos) :
+		_stockPileRender(cardAtlas),
+		_emptyWastePile(klondikeAtlas),
 		_cardAtlas(&cardAtlas),
 		_klondikeAtlas(&klondikeAtlas),
+		_font(font),
 		_startPosition(startPos)
 	{
 		setKlondikeField(field);
@@ -28,6 +16,30 @@ namespace WGP {
 		setColumnSpacing(colSpacing);
 		setRowSpacing(rowSpacing);
 		setRendering(true);
+
+		stockPileSpriteInit();
+		_stockPileRender.setPosition(_startPosition);
+
+		remainingCardsSetText();
+		remainingCardsTextInit();
+
+		sf::Vector2f wastePos = startPos + sf::Vector2f(colSpacing, 0.f);
+		emptyWastePileSpriteInit();
+		_emptyWastePile.setPosition(wastePos);
+
+		_wastePileRender.initializeFromPosition(cardAtlas, &(_klondikeField->getWastePile()), wastePos, wastePos, scale, sf::Angle(sf::degrees(0)));
+
+		TVector<sf::Vector2f> coords = getCoordinatesVectorFromAngle(4, wastePos + sf::Vector2f(2 * colSpacing, 0.f), sf::Angle(sf::degrees(0)), colSpacing);
+		emptyFoundationPilesSpritesInit();
+		for (int i = 0; i < coords.size(); i++) {
+			_emptyFoundationPiles[i].setPosition(coords[i]);
+			_foundationPilesRenders[i].initializeFromPosition(cardAtlas, &(_klondikeField->getFoundationPile(static_cast<CardSuit>(i))), coords[i], coords[i], scale, sf::Angle(sf::degrees(0)));
+		}
+
+		coords = getCoordinatesVectorFromAngle(7, startPos + sf::Vector2f(0.f, rowSpacing), sf::Angle(sf::degrees(0)), colSpacing);
+		for (int i = 0; i < coords.size(); i++) {
+			_tableauPilesRenders[i].initializeFromAngle(cardAtlas, &(_klondikeField->getTableauPile(i)), coords[i], sf::Angle(sf::degrees(90)), cardsSpacing, scale, sf::Angle(sf::degrees(0)));
+		}
 	}
 
 	// Setters
@@ -41,29 +53,32 @@ namespace WGP {
 	}
 
 	void klondikeFieldRenderer::setCardAtlas(sf::Texture& cardAtlas) {
+		// Мб переписать
 		_cardAtlas = &cardAtlas;
-		//for (int i = 0; i < 7; i++) _tableauPilesRenders[i].setCardAtlas(cardAtlas);
-		//for (int i = 0; i < 4; i++) _foundationPilesRenders[i].setCardAtlas(cardAtlas);
-		//_stockPileRender.setCardAtlas(cardAtlas);
-		//_wastePileRender.setCardAtlas(cardAtlas);
-		//update()
+		for (int i = 0; i < 7; i++) _tableauPilesRenders[i].setCardAtlas(cardAtlas);
+		for (int i = 0; i < 4; i++) _foundationPilesRenders[i].setCardAtlas(cardAtlas);
+		_stockPileRender.setTexture(*_cardAtlas);
+		_stockPileRender.setTextureRect(coordsToIntRectCardAtlas(sf::Vector2u(0, 4)));
+		_wastePileRender.setCardAtlas(cardAtlas);
+		//update();
 	}
 
 	void klondikeFieldRenderer::setKlondikeAtlas(sf::Texture& klondikeAtlas) {
 		_klondikeAtlas = &klondikeAtlas;
-		// ???
-		//update()
+		//emptyStockPileSpriteInit();
+		//emptyFoundationPilesSpritesInit();
 	}
 
 	void klondikeFieldRenderer::setScale(float scale) {
+		// Мб переписать
+		_cardsScale = scale;
 		if (scale <= 0) {
 			throw std::invalid_argument("klondikeFieldRenderer.setScale: Invalid argument 'scale' - must be > 0");
 		}
-		for (int i = 0; i < 7; i++) _tableauPilesRenders[i].setCardsScale(scale);
+		/*for (int i = 0; i < 7; i++) _tableauPilesRenders[i].setCardsScale(scale);
 		for (int i = 0; i < 4; i++) _foundationPilesRenders[i].setCardsScale(scale);
 		_stockPileRender.setCardsScale(scale);
-		_wastePileRender.setCardsScale(scale);
-		//update()
+		_wastePileRender.setCardsScale(scale);*/
 	}
 
 	void klondikeFieldRenderer::setCardsSpacing(float cardsSpacing) {
@@ -71,7 +86,7 @@ namespace WGP {
 			throw std::invalid_argument("klondikeFieldRenderer.setCardsSpacing: Invalid argument 'cardsSpacing' - must be > 0");
 		}
 		_cardsSpacing = cardsSpacing;
-		//update()
+		//update();
 	}
 
 	void klondikeFieldRenderer::setColumnSpacing(float colSpacing) {
@@ -79,7 +94,7 @@ namespace WGP {
 			throw std::invalid_argument("klondikeFieldRenderer.setColumnSpacing: Invalid argument 'colSpacing' - must be > 0");
 		}
 		_columnSpacing = colSpacing;
-		//update()
+		//update();
 	}
 
 	void klondikeFieldRenderer::setRowSpacing(float rowSpacing) {
@@ -87,12 +102,12 @@ namespace WGP {
 			throw std::invalid_argument("klondikeFieldRenderer.setRowSpacing: Invalid argument 'rowSpacing' - must be > 0");
 		}
 		_rowSpacing = rowSpacing;
-		//update()
+		//update();
 	}
 
 	void klondikeFieldRenderer::setStartPosition(sf::Vector2f startPos) {
 		_startPosition = startPos;
-		//update()
+		//update();
 	}
 
 	void klondikeFieldRenderer::setRendering(bool isRendering) {
@@ -101,16 +116,40 @@ namespace WGP {
 
 	// Functions
 
-	void klondikeFieldRenderer::draw(sf::RenderTarget&) {
-
+	void klondikeFieldRenderer::draw(sf::RenderTarget& target) {
+		// Разделить draw и update
+		if (!_isRendering) return;
+		if (!_klondikeField->getStockPile().isEmpty()) {
+			target.draw(_stockPileRender);
+		}
+		remainingCardsSetText();
+		target.draw(_remainingCardsInStock);
+		if (_klondikeField->getWastePile().isEmpty()) {
+			target.draw(_emptyWastePile);
+		}
+		_wastePileRender.draw(target);
+		for (int i = 0; i < 4; i++) {
+			if (_klondikeField->getFoundationPile(static_cast<CardSuit>(i)).isEmpty()) {
+				target.draw(_emptyFoundationPiles[i]);
+			}
+			_foundationPilesRenders[i].draw(target);
+		}
+		for (int i = 0; i < 7; i++) {
+			_tableauPilesRenders[i].draw(target);
+		}
 	}
 
 
 	void klondikeFieldRenderer::update() {
+		// Переписать
 		for (int i = 0; i < 7; i++) {
 			_tableauPilesRenders[i].update();
 		}
-		// ??
+		for (int i = 0; i < 4; i++) {
+			_foundationPilesRenders[i].update();
+		}
+		//_stockPileRender.update();
+		_wastePileRender.update();
 	}
 
 	// Getters
@@ -137,5 +176,73 @@ namespace WGP {
 
 	bool klondikeFieldRenderer::isRendering() const {
 		return _isRendering;
+	}
+
+	// Private function
+
+	void klondikeFieldRenderer::stockPileSpriteInit() {
+		_stockPileRender.setTexture(*_cardAtlas);
+		_stockPileRender.setTextureRect(coordsToIntRectCardAtlas(sf::Vector2u(0, 4)));
+		_stockPileRender.setOrigin(getLocalCentreOfSprite(_stockPileRender));
+		_stockPileRender.setRotation(sf::Angle(sf::degrees(0)));
+		_stockPileRender.setScale(sf::Vector2f(_cardsScale, _cardsScale));
+	}
+
+	void klondikeFieldRenderer::emptyWastePileSpriteInit() {
+		_emptyWastePile.setTexture(*_klondikeAtlas);
+		_emptyWastePile.setTextureRect(coordsToIntRectKlondikeAtlas(0));
+		_emptyWastePile.setOrigin(getLocalCentreOfSprite(_emptyWastePile));
+		_emptyWastePile.setRotation(sf::Angle(sf::degrees(0)));
+		_emptyWastePile.setScale(sf::Vector2f(_cardsScale, _cardsScale));
+
+	}
+
+	void klondikeFieldRenderer::emptyFoundationPilesSpritesInit() {
+		for (int i = 0; i < 4; i++) {
+			_emptyFoundationPiles[i].setTexture(*_klondikeAtlas);
+			_emptyFoundationPiles[i].setTextureRect(coordsToIntRectKlondikeAtlas(1));
+			_emptyFoundationPiles[i].setOrigin(getLocalCentreOfSprite(_emptyWastePile));
+			_emptyFoundationPiles[i].setRotation(sf::Angle(sf::degrees(0)));
+			_emptyFoundationPiles[i].setScale(sf::Vector2f(_cardsScale, _cardsScale));
+		}
+	}
+
+	void klondikeFieldRenderer::remainingCardsTextInit() {
+		_remainingCardsInStock.setFont(_font);
+		_remainingCardsInStock.setFillColor(sf::Color(0, 0, 0, 255));
+		_remainingCardsInStock.setOrigin(getLocalCentreOfText(_remainingCardsInStock));
+		_remainingCardsInStock.setScale(sf::Vector2f(_cardsScale, _cardsScale));
+		_remainingCardsInStock.setPosition(_startPosition);
+	}
+
+	void klondikeFieldRenderer::remainingCardsSetText() {
+		_remainingCardsInStock.setString(std::to_string(_klondikeField->getStockPile().topIndex() + 1));
+		_remainingCardsInStock.setOrigin(getLocalCentreOfText(_remainingCardsInStock));
+	}
+
+	// Supporitive functions
+
+	sf::IntRect coordsToIntRectKlondikeAtlas(unsigned int index) {
+		sf::Vector2i size(CARD_SPRITE_WIDTH, CARD_SPRITE_HEIGHT);
+		int posX = 0, posY = 0;
+		if (index != 0 && index != 1) {
+			throw std::out_of_range("coordscoordsToIntRectKlondikeAtlas: Invalid argument 'index' - coords out of range");
+		}
+		else {
+			posX = CARD_ATLAS_SPACING + index * (CARD_SPRITE_WIDTH + CARD_ATLAS_SPACING);
+			posY = CARD_ATLAS_SPACING + 0 * (CARD_SPRITE_HEIGHT + CARD_ATLAS_SPACING);
+		}
+		sf::Vector2i pos(posX, posY);
+		return sf::IntRect(pos, size);
+	}
+
+	sf::Vector2f getLocalCentreOfText(sf::Text& text) {
+		sf::FloatRect bounds = text.getLocalBounds();
+		return sf::Vector2f(bounds.getCenter());
+	}
+
+	sf::Vector2f getGlobalCentreOfText(sf::Text& text) {
+		sf::FloatRect bounds = text.getGlobalBounds();
+		return sf::Vector2f(bounds.getCenter());
 	}
 }