+++ src/games/solitaire/klondike/klondikeFieldRenderer/klondikeFieldRenderer.h
@@ -4,7 +4,6 @@
 #include "../../../../core/fundamental classes/render/cardRenderer/cardRenderer.h"
 #include "../../../../core/fundamental classes/render/deckRenderer/deckRenderer.h"
 #include "../../../../core/fundamental classes/TVector/TVector.h"
-#include "../../../../main/constants.h"
 #include <iostream>
 
 #pragma once
@@ -15,12 +14,17 @@ namespace WGP {
 		klondikeField* _klondikeField;
 		sf::Texture* _cardAtlas;
 		sf::Texture* _klondikeAtlas;
+		const sf::Font _font;
 		float _cardsScale;
 
-		TVector<deckRenderer> _tableauPilesRenders{ 7 }; // 7
-		TVector<deckRenderer> _foundationPilesRenders{ 4 }; // 4
-		deckRenderer _stockPileRender;
+		TVector<deckRenderer> _tableauPilesRenders{ 7 };
+		TVector<deckRenderer> _foundationPilesRenders{ 4 };
+		sf::Sprite _stockPileRender;
+		sf::Text _remainingCardsInStock{ _font };
 		deckRenderer _wastePileRender;
+
+		sf::Sprite _emptyWastePile;
+		sf::Sprite _emptyFoundationPiles[4] = { sf::Sprite(*_klondikeAtlas), sf::Sprite(*_klondikeAtlas), sf::Sprite(*_klondikeAtlas), sf::Sprite(*_klondikeAtlas) };
 
 		float _cardsSpacing;
 		float _columnSpacing;
@@ -30,8 +34,7 @@ namespace WGP {
 		bool _isRendering;
 	public:
 		// Constructors
-		klondikeFieldRenderer();
-		klondikeFieldRenderer(sf::Texture&, sf::Texture&, klondikeField*, float, float, float, float, sf::Vector2f);
+		klondikeFieldRenderer(sf::Texture&, sf::Texture&, klondikeField*, const sf::Font&, float, float, float, float, sf::Vector2f);
 
 		// Setters
 		void setKlondikeField(klondikeField*);
@@ -55,5 +58,15 @@ namespace WGP {
 		float getRowSpacing() const;
 		sf::Vector2f getStartPosition() const;
 		bool isRendering() const;
+	private:
+		void stockPileSpriteInit();
+		void emptyWastePileSpriteInit();
+		void emptyFoundationPilesSpritesInit();
+		void remainingCardsTextInit();
+		void remainingCardsSetText();
 	};
+
+	sf::IntRect coordsToIntRectKlondikeAtlas(unsigned int);
+	sf::Vector2f getLocalCentreOfText(sf::Text&);
+	sf::Vector2f getGlobalCentreOfText(sf::Text&);
 }